Deployment tool
This adds bundle-deploy to the tools installed on the bundle-gen
image. It can equally be built and used without being inside the
image.

The main problem this has to solve is that the store accepts only
monotonically incrementing positive integers as version numbers, but
these make it quite hard to build test or homebrew bundles, and these
values are meaningless to users. In a CI environment, it's often
possible to obtain this kind of version number automatically, but in
other environments it can also require more work.

This takes a .bundle file, as emitted by bundle-gen, and converts it
into a ZIP file with the provided version number. The old version
number from the bundle is sent to the store as the "displayVersion",
so that whatever the .bundle file had as its version will be what
users see on the store page. The ZIP file is preserved, and can be
stored separately as a CI artifact if needed.

bundle-gen emits .bundle files, because they are easy to capture in
patterns in scripts, and there is minimal chance of accidentally
picking up downloaded ZIP files from the build process. The version is
replaced in the file uploaded to the store so that bundle-gen remains
suitable for Homebrew, where versioning needs to be human readable
(e.g. semver). This is also a convenient way for the build process run
by bundle-gen to communicate the display version to the store upload.