Query tool for reading bundle fields in scripts
This is a simple tool for reading back the values in bundle.ini from a
ZIP file, without resorting to grep, or other similar approaches. Pass
-f to read a particular field; if the field has multiple values they
will be emitted one per line, which is convenient for scripting. The
exit status will be 0 if the requested field was present, and 1 if it
was not. Note that Booleans in bundle.ini that match their default
setting will also show as not present; the two cases are
indistinguishable for the tool (not present vs default value).

As a sanity check, you can pass -a to get more human readable output
of all of the fields, as parsed by the tool, suitable for logging.